Joined together BoxedStrict and BoxedNF modules
