+++ massiv/src/Data/Massiv/Array/Manifest.hs
@@ -34,8 +34,7 @@ module Data.Massiv.Array.Manifest
   , Unbox
   ) where
 
-import           Data.Massiv.Array.Manifest.BoxedStrict
-import           Data.Massiv.Array.Manifest.BoxedNF
+import           Data.Massiv.Array.Manifest.Boxed
 import           Data.Massiv.Array.Manifest.Internal
 import           Data.Massiv.Array.Manifest.Primitive
 import           Data.Massiv.Array.Manifest.Storable
